+++ src/test_socket_can_node.cpp
@@ -25,7 +25,7 @@ int main(int argc, char **argv)
     ros::Rate r(rate);
 	const char *interface_name = "vcan0";
     CANDriver *can_driver = new CANDriver(interface_name); // speed controller instance: It includes all the functions that perform PID feedback control
-	int data [2] = {0x11,0x22};
+	uint8_t data [2] = {0x11,0x22};
     while (nh.ok())
     {
 		can_driver->CANWrite(0x123,2,data,2);
